Move tests for the `MapStringifier` to the `MapStringifierShould` class.
